web_app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-multipart
websockets
redis
neo4j
httpx
PyJWT[crypto]
passlib[bcrypt]
python-socketio
aiofiles
jinja2
torch==2.2.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
scikit-learn
transformers
sentence-transformers
openai
dashscope
asyncio-mqtt
loguru
rich
typer
click
spacy
